Reverting to int form

+++ tempsensor_v1/src/config.h
@@ -434,7 +434,7 @@ typedef struct {
 	uint32_t lastSeek;
 
 	// GSM or GPRS
-	uint8_t network_mode;
+	int network_mode;
 	NETWORK_SERVICE net_service[2];
 
 	SAFEBOOT_STATUS safeboot;
